# Conflicts:
#	USER_GUIDE.md
#	lib/Epub/Epub/BookMetadataCache.cpp
#	lib/Epub/Epub/Section.cpp
#	lib/Epub/Epub/Section.h
#	lib/I18n/translations/english.yaml
#	lib/I18n/translations/russian.yaml
#	lib/KOReaderSync/ProgressMapper.cpp
#	lib/KOReaderSync/ProgressMapper.h
#	src/JsonSettingsIO.cpp
#	src/JsonSettingsIO.h
#	src/activities/ActivityResult.h
#	src/activities/reader/EpubReaderActivity.cpp
#	src/activities/reader/EpubReaderActivity.h
#	src/activities/reader/EpubReaderMenuActivity.h
#	src/activities/reader/KOReaderSyncActivity.cpp
#	src/activities/reader/KOReaderSyncActivity.h
#	src/activities/reader/ReaderUtils.h
#	src/components/themes/BaseTheme.h
#	src/components/themes/lyra/LyraTheme.cpp

+++ src/activities/home/RecentBooksActivity.cpp
@@ -15,7 +15,7 @@
 
 namespace {
 // Hold threshold for the long-press "remove from list" action (firmware convention).
-constexpr unsigned long LONG_PRESS_MS = 1000;
+constexpr uint32_t LONG_PRESS_MS = 1000;
 }  // namespace
 
 void RecentBooksActivity::loadRecentBooks() { recentBooks = RECENT_BOOKS.getBooks(); }
